finished code splitting the files
TODO: update docs
