+++ src/getCurrentUA.ts
@@ -1,0 +1,7 @@
+﻿let getCurrentUA = (): string => {
+    return navigator.userAgent;
+};
+let currentUA = getCurrentUA();
+
+
+export { currentUA, getCurrentUA }
