+++ src/index.ts
@@ -9,63 +9,21 @@ Licensed under the Apache-2.0 License.
 
 import { DetectScreenOrientation, ORIENTATION_isLandscape } from './orientation'
 
+// Device Type
 
-//DEVICE TYPE
-function DetectDeviceType(): "tablet" | "mobile" | "desktop" {
-    const ua = navigator.userAgent;
-    if (/(tablet|ipad|playbook|silk)|(android(?!.*mobi))/i.test(ua)) {
-        return "tablet";
-    }
-    else if (/(iPhone|iPod|iPad)/i.test(ua) || /(android)/i.test(ua) || /(windows phone)/i.test(ua) || /(blackberry)/i.test(ua) || /(bb10)/i.test(ua) || /(opera mini)/i.test(ua) || /(opera mobi)/i.test(ua) || /(iemobile)/i.test(ua) || /(symbian)/i.test(ua) || /(maemo)/i.test(ua) || /(webos)/i.test(ua) || /(mobile)/i.test(ua) || /(tablet)/i.test(ua) || /(symbianos)/i.test(ua) || /(up.browser)/i.test(ua) || /(up.link)/i.test(ua) || /(mmp)/i.test(ua) || /(symbianos)/i.test(ua) || /(midp)/i.test(ua) || /(wap)/i.test(ua) || /(phone)/i.test(ua) || /(pocket)/i.test(ua) || /(mobile)/i.test(ua) || /(pda)/i.test(ua) || /(avantgo)/i.test(ua) || /(eudoraweb)/i.test(ua) || /(brew)|(bada)/i.test(ua) || /(blackberry)/i.test(ua) || /(hpwos)/i.test(ua) || /(kindle)/i.test(ua) || /(lge.netcast)/i.test(ua) || /(lg;lx)|(lg;lx)/i.test(ua) || /(lge.netcast)/i.test(ua) || /(nintendo.3ds)/i.test(ua) || /(nintendo.ds)/i.test(ua) || /(nintendo.wiiu)/i.test(ua) || /(nintendo.wii)/i.test(ua) || /(playstation.3ds)/i.test(ua) || /(playstation.ds)/i.test(ua) || /(playstation.wiiu)/i.test(ua) || /(playstation.wii)/i.test(ua) || /(webos)/i.test(ua)) {
-        return "mobile";
-        // add "| /(firefox)/i.test(ua) |" if you want to detect Firefox (not tested for mobile)
-    }
-    else {
-        return "desktop";
-    }
+import { DEVICE_type, DetectDeviceType } from './DetectDeviceType'
 
-}
-let DEVICE_type = DetectDeviceType();
+// Finite Mobile Device Type
 
-function finiteMobileDeviceType(): "iOS" | "Android" | "BlackBerry" | "Windows Phone" | "webOS" | "unknown" {
-    const ua = navigator.userAgent;
-    if (/(iPhone|iPod|iPad)/i.test(ua)) {
-        console.log("iOS");
-        return "iOS";
-    }
-    else if (/Android/i.test(ua)) {
-        console.log("Android");
-        return "Android";
-    }
-    else if (/BlackBerry/i.test(ua)) {
-        console.log("BlackBerry");
-        return "BlackBerry";
-    }
-    else if (/Windows Phone/i.test(ua)) {
-        console.log("Windows Phone");
-        return "Windows Phone";
-    }
-    else if (/webOS/i.test(ua)) {
-        console.log("webOS");
-        return "webOS";
-    }
-    else {
-        console.log("ERR: Unknown Device");
-        return "unknown";
-    }
-}
+import { DEVICE_finiteType, finiteMobileDeviceType } from './finiteMobileDeviceType'
 
-let DEVICE_finiteType = finiteMobileDeviceType();
+// standard UA
 
 
-let getCurrentUA = (): string => {
-    return navigator.userAgent;
-}
-
-let currentUA = getCurrentUA();
-
+import { getCurrentUA, currentUA } from './getCurrentUA'
 
 // Exports
+
 
 
 // Make the functions accessible
