+++ tsconfig.json
@@ -4,7 +4,7 @@
         "outDir": "./"
     },
     "files": [
-        "./src/index.ts"
+        "./src/index.ts", "./src/getCurrentUA.ts", "./src/finiteMobileDeviceType.ts", "./src/DetectDeviceType.ts"
     ],
     "types": [
         "jquery",
